Add GraphApi::Common and GraphApi::Energy
Adds support for separate APIs depending on the purpose of the graph.
